+++ scripts/gpsans_reduction.py
@@ -11,7 +11,7 @@ import mantid.simpleapi as msapi  # noqa E402
 import drtsans  # noqa E402
 from drtsans.mono import gpsans as sans  # noqa E402
 from drtsans.settings import unique_workspace_dundername as uwd  # noqa E402
-
+from drtsans.path import registered_workspace # noqa #402
 from common_utils import get_Iq, get_Iqxqy, setup_configuration  # noqa E402
 
 INSTRUMENT = 'GPSANS'
@@ -54,11 +54,21 @@ def reduction(json_params, config):
     """
         Perform the whole reduction
     """
+    sensitivity_workspace = None
     sensitivity_file_path = config['sensitivity_file_path']
-    config.pop('sensitivity_file_path')
-    sensitivity_workspace = uwd()
-    drtsans.load_sensitivity_workspace(sensitivity_file_path, sensitivity_workspace)
-    config['sensitivity_workspace'] = sensitivity_workspace
+    if sensitivity_file_path is not None:
+        config.pop('sensitivity_file_path')
+        sensitivity_workspace = uwd()
+        drtsans.load_sensitivity_workspace(sensitivity_file_path, sensitivity_workspace)
+        config['sensitivity_workspace'] = sensitivity_workspace
+
+    dark_current_workspace = None
+    dark_current_file_path = config['dark_current']
+    if dark_current_file_path is not None:
+        dark_current_workspace = uwd()
+        sans.load_dark_current_workspace(dark_current_file_path, dark_current_workspace)
+        config['dark_current'] = dark_current_workspace
+
     # Load and prepare scattering data
     ws = sans.prepare_data(json_params["instrumentName"] + '_' + json_params["runNumber"],
                            output_suffix='_data', **config)
@@ -86,7 +96,11 @@ def reduction(json_params, config):
         # Subtract background
         ws = drtsans.subtract_background(ws, background=ws_bck)
         msapi.logger.notice("Background subtracted")
-    msapi.DeleteWorkspace(sensitivity_workspace)
+
+    if registered_workspace(sensitivity_workspace):
+        msapi.DeleteWorkspace(sensitivity_workspace)
+    if registered_workspace(dark_current_workspace):
+        msapi.DeleteWorkspace(dark_current_workspace)
 
     # Final normalization
     try:
